+++ internal/azure/types/integer_type.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 
 	"github.com/Azure/terraform-provider-azapi/internal/azure/utils"
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
 var _ TypeBase = &IntegerType{}
@@ -26,29 +28,29 @@ func (t *IntegerType) AsTypeBase() *TypeBase {
 	return &typeBase
 }
 
-func (t *IntegerType) Validate(body interface{}, path string) []error {
-	if body == nil {
+func (t *IntegerType) Validate(body attr.Value, path string) []error {
+	if body == nil || body.IsNull() || body.IsUnknown() {
 		return nil
 	}
-	var v int
+	var v int64
 	switch input := body.(type) {
-	case float64, float32:
+	case types.Int32:
+		v = int64(input.ValueInt32())
+	case types.Int64:
+		v = input.ValueInt64()
+	case types.Float64, types.Float32, types.Number:
 		// TODO: skip validation for now because of the following issue:
 		// the bicep-types-az parses float as integer type and it should be fixed: https://github.com/Azure/bicep-types-az/issues/1404
 		return nil
-	case int64:
-		v = int(input)
-	case int32:
-		v = int(input)
-	case int:
-		v = input
+	case types.Dynamic:
+		return t.Validate(input.UnderlyingValue(), path)
 	default:
 		return []error{utils.ErrorMismatch(path, "integer", fmt.Sprintf("%T", body))}
 	}
-	if t.MinValue != nil && v < *t.MinValue {
+	if t.MinValue != nil && v < int64(*t.MinValue) {
 		return []error{utils.ErrorCommon(path, fmt.Sprintf("value is less than %d", *t.MinValue))}
 	}
-	if t.MaxValue != nil && v > *t.MaxValue {
+	if t.MaxValue != nil && v > int64(*t.MaxValue) {
 		return []error{utils.ErrorCommon(path, fmt.Sprintf("value is greater than %d", *t.MaxValue))}
 	}
 	return nil