+++ internal/azure/types/string_literal_type.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 
 	"github.com/Azure/terraform-provider-azapi/internal/azure/utils"
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
 var _ TypeBase = &StringLiteralType{}
@@ -27,16 +29,20 @@ func (t *StringLiteralType) GetWriteOnly(i interface{}) interface{} {
 	return i
 }
 
-func (t *StringLiteralType) Validate(body interface{}, path string) []error {
-	if t == nil || body == nil {
+func (t *StringLiteralType) Validate(body attr.Value, path string) []error {
+	if t == nil || body == nil || body.IsNull() || body.IsUnknown() {
 		return []error{}
 	}
 	errors := make([]error, 0)
-	if stringValue, ok := body.(string); ok {
-		if stringValue != t.Value {
-			errors = append(errors, utils.ErrorMismatch(path, t.Value, stringValue))
+
+	switch v := body.(type) {
+	case types.String:
+		if v.ValueString() != t.Value {
+			errors = append(errors, utils.ErrorMismatch(path, t.Value, v.ValueString()))
 		}
-	} else {
+	case types.Dynamic:
+		return t.Validate(v.UnderlyingValue(), path)
+	default:
 		errors = append(errors, utils.ErrorMismatch(path, "string", fmt.Sprintf("%T", body)))
 	}
 	return errors