+++ internal/services/azapi_data_plane_resource_test.go
@@ -369,10 +369,6 @@ resource "azapi_data_plane_resource" "test" {
     }
   }
 
-  retry = {
-    error_message_regex = ["403 Forbidden"]
-  }
-
   depends_on = [
     azapi_resource_action.add_accesspolicy
   ]