+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v2.4.0 (unreleased)
 
+ENHANCEMENTS:
+- `azapi_resource` resource: Support schema validation in `terraform validate`.
+
 BUG FIXES:
 - Fix a bug that `azapi_update_resource` resource produced inconsistent results when only `error_message_regex` is changed.
 - Fix a bug that `azapi_resource_action` resource could not be migrated correctly when the `body` is empty string.